backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
scikit-learn
statsmodels
ta
numpy
pytest
pytest-mock
httpx